hpo: move splits to training, harden the baseline guard, seed-parity folds, one trial-record schema
