+++ src/workbench/model_scripts/chemprop/chemprop.template
@@ -372,7 +372,6 @@ if __name__ == "__main__":
         check_dataframe,
         compute_classification_metrics,
         compute_regression_metrics,
-        get_split_indices,
         print_classification_metrics,
         print_confusion_matrix,
         print_regression_metrics,
@@ -454,6 +453,7 @@ if __name__ == "__main__":
 
     # Route held-out validation rows out of training BEFORE label encoding / feature-mean fitting,
     # so nothing downstream (CV/UQ/SHAP) sees them. They're scored held-out just before save.
+    from workbench.training.splits import get_split_indices
     from workbench.training.validation import split_validation_set
 
     all_df, holdout_df = split_validation_set(all_df)
@@ -601,7 +601,7 @@ if __name__ == "__main__":
         strategy=split_strategy,
         target_column=target_col,
         test_size=0.2,
-        random_state=42,
+        random_state=hyperparameters["seed"],
         butina_cutoff=butina_cutoff,
     )
     print(f"Split strategy: {split_strategy}")
